requirements.txt
torch==2.5.1 -i https://download.pytorch.org/whl/cu118
torchvision==0.20.1 -i https://download.pytorch.org/whl/cu118
torchaudio==2.5.1 -i https://download.pytorch.org/whl/cu118
ftfy
packaging
regex
tqdm
albumentations==1.4.2
numpy==1.26.4
matplotlib
pandas
opencv-python
wandb
pyarrow==21.0.0
torch_geometric==2.3.1
hydra-core==1.3.2
omegaconf==2.3.0
openTSNE
pytorch_lightning==2.0.4
myopic-mces --no-deps
pulp>=2.7.0
lmdb